clobber_aosp: Only do clean builds if we are told too
Weird errors occurs that require restarting the build. By default don't
clean but give the user the option to force clean builds if they want
to.

+++ vars/clobber_aosp.groovy
@@ -1,11 +1,13 @@
-def call() {
+def call(clean_build) {
 	sh """
-		if [ -d aosp ]; then
-			pushd aosp
-			source build/envsetup.sh
-			m clean
-			repo forall -c git clean -xdf
-			popd
+		if [ "${clean_build}" == "true" ]; then
+			if [ -d aosp ]; then
+				pushd aosp
+				source build/envsetup.sh
+				m clean
+				repo forall -c git clean -xdf
+				popd
+			fi
 		fi
 	"""
 }
